Remove long double from primitive types

+++ modules/ballerina-core/src/main/java/org/ballerinalang/bre/BLangExecutor.java
@@ -417,7 +417,7 @@ public class BLangExecutor implements NodeExecutor {
         BMessage inMsg = (BMessage) expr.execute(this);
         List<WorkerRunner> workerRunnerList = new ArrayList<>();
         List<BMessage> resultMsgs = new ArrayList<>();
-        int timeout = ((BInteger) forkJoinStmt.getTimeout().getTimeoutExpression().execute(this)).intValue();
+        long timeout = ((BInteger) forkJoinStmt.getTimeout().getTimeoutExpression().execute(this)).intValue();
 
         Worker[] workers = forkJoinStmt.getWorkers();
         Map<String, WorkerRunner> triggeredWorkers = new HashMap<>();
@@ -526,7 +526,7 @@ public class BLangExecutor implements NodeExecutor {
 
     }
 
-    private BMessage invokeAnyWorker(List<WorkerRunner> workerRunnerList, int timeout) {
+    private BMessage invokeAnyWorker(List<WorkerRunner> workerRunnerList, long timeout) {
         ExecutorService anyExecutor = Executors.newWorkStealingPool();
         BMessage result;
         try {
@@ -540,7 +540,7 @@ public class BLangExecutor implements NodeExecutor {
         return result;
     }
 
-    private List<BMessage> invokeAllWorkers(List<WorkerRunner> workerRunnerList, int timeout) {
+    private List<BMessage> invokeAllWorkers(List<WorkerRunner> workerRunnerList, long timeout) {
         ExecutorService allExecutor = Executors.newWorkStealingPool();
         List<BMessage> result = new ArrayList<>();
         try {
@@ -740,7 +740,8 @@ public class BLangExecutor implements NodeExecutor {
             if (arrayMapAccessExpr.getType() != BTypes.typeMap) {
                 // Get the value stored in the index
                 if (collectionValue instanceof BArray) {
-                    return ((BArray) collectionValue).get(((BInteger) indexValue).intValue());
+                    //Here we cast index value to int as anyway java doesn't support long as array indexes
+                    return ((BArray) collectionValue).get((int) ((BInteger) indexValue).intValue());
                 } else {
                     return collectionValue;
                 }
@@ -1014,7 +1015,8 @@ public class BLangExecutor implements NodeExecutor {
         if (!(accessExpr.getType() == BTypes.typeMap)) {
             BArray arrayVal = (BArray) accessExpr.getRExpr().execute(this);
             BInteger indexVal = (BInteger) accessExpr.getIndexExpr().execute(this);
-            arrayVal.add(indexVal.intValue(), rValue);
+            //Here we cast index value to int as anyway java doesn't support long as array indexes
+            arrayVal.add((int) indexVal.intValue(), rValue);
 
         } else {
             BMap<BString, BValue> mapVal = (BMap<BString, BValue>) accessExpr.getRExpr().execute(this);
@@ -1199,7 +1201,8 @@ public class BLangExecutor implements NodeExecutor {
         if (fieldExpr.getRefVarType() instanceof BMapType) {
             ((BMap) arrayMapValue).put(indexValue, rValue);
         } else {
-            ((BArray) arrayMapValue).add(((BInteger) indexValue).intValue(), rValue);
+            //Here we cast index value to int as anyway java doesn't support long as array indexes
+            ((BArray) arrayMapValue).add((int) ((BInteger) indexValue).intValue(), rValue);
         }
     }
 
@@ -1272,7 +1275,8 @@ public class BLangExecutor implements NodeExecutor {
         if (fieldExpr.getRefVarType() instanceof BMapType) {
             unitVal = ((BMap) currentVal).get(indexValue);
         } else {
-            unitVal = ((BArray) currentVal).get(((BInteger) indexValue).intValue());
+            //Here we cast index value to int as anyway java doesn't support long as array indexes
+            unitVal = ((BArray) currentVal).get((int) ((BInteger) indexValue).intValue());
         }
 
         if (unitVal == null) {
