+++ modules/ballerina-core/src/test/java/org/ballerinalang/model/expressions/SimpleTypeCastExprTest.java
@@ -80,7 +80,7 @@ public class SimpleTypeCastExprTest {
         testFloatToBooleanCast(0, false);
     }
 
-    private void testBooleanToIntCast(Boolean input, int excepted, String functionName) {
+    private void testBooleanToIntCast(Boolean input, long excepted, String functionName) {
         BValue[] args = { new BBoolean(input) };
         BValue[] returns = BLangFunctions.invoke(bLangProgram, functionName, args);
         Assert.assertEquals(returns.length, 1);
@@ -88,7 +88,7 @@ public class SimpleTypeCastExprTest {
         Assert.assertEquals(((BInteger) returns[0]).intValue(), excepted);
     }
 
-    private void testBooleanToFloatCast(Boolean input, float excepted, String functionName) {
+    private void testBooleanToFloatCast(Boolean input, double excepted, String functionName) {
         BValue[] args = { new BBoolean(input) };
         BValue[] returns = BLangFunctions.invoke(bLangProgram, functionName, args);
         Assert.assertEquals(returns.length, 1);
