+++ modules/ballerina-native/src/main/java/org/ballerinalang/nativeimpl/lang/datatables/GetValueAsStringByIndex.java
@@ -57,7 +57,8 @@ public class GetValueAsStringByIndex extends AbstractNativeFunction {
 
     public BValue[] execute(Context ctx) {
         BDataTable dataTable = (BDataTable) getArgument(ctx, 0);
-        int index = ((BInteger) getArgument(ctx, 1)).intValue();
+        //Here we cast index value to int for consistency
+        int index = (int) ((BInteger) getArgument(ctx, 1)).intValue();
         return getBValues(new BString(dataTable.getObjectAsString(index)));
     }
 }
